deploy_temp/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
wechat-python>=0.0.2
itchat>=1.3.10
schedule>=1.2.0
python-dotenv>=1.0.0
jaydebeapi>=1.2.3
fake-useragent>=1.4.0
paramiko>=3.0.0
